requirements.txt
ultralytics>=8.3.0
torch>=2.1.0
torchvision>=0.16.0
albumentations>=1.4.0
opencv-python>=4.9.0
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
wandb>=0.17.0
tqdm>=4.66.0
Pillow>=10.2.0
grad-cam>=1.5.0
onnx>=1.16.0
onnxruntime>=1.18.0
